refactor: Lombok, Jackson error 직렬화 문제로 인해 수정

+++ src/main/java/com/carbonhater/co2zerobookmark/bookmark/model/dto/FolderHierarchyDto.java
@@ -45,7 +45,7 @@ public class FolderHierarchyDto {
         private String bookmarkUrl;
 
         public BookmarkDto(Bookmark bookmark) {
-            this.bookmarkId = bookmark.getBookmarkId();
+            this.bookmarkId = (Long) bookmark.getBookmarkId();
             this.bookmarkName = bookmark.getBookmarkName();
             this.bookmarkUrl = bookmark.getBookmarkUrl();
         }
